Split resource voting routes into separate module
