+++ app/api/routes/__init__.py
@@ -1,4 +1,5 @@
 from app.api.routes import old_routes
 from app.api.routes import resources
+from app.api.routes import resource_voting
 from app.api.routes import categories
 from app.api.routes import languages
